Start of the Dependency Resolution System.
When the Dependency Res System object is created, filters are added(64/32,max vers), and PkgAgents that need to be resolved.  These PkgAgents have a sorted list of its packages available and then add additional PkgAgents with constraints. Then we iteratively update the tree that we have made from these PkgAgents until they don't change anymore... there is much more to it..read comments in code for a better explanation as it comes in...